feat(strawberry): Use operation name as transaction name (#3294)
The Strawberry integration is creating spans at the moment, but they're all grouped under the same /graphql transaction coming from the web framework integration. This has significant effect on the usefulness of tracing.

With this change we start using the operation name to update the name of the transaction so that each unique operation becomes its own event group.