Make a deeper copy of the vars object with fallback to safe_repr
To make this work, I had to add a port of python's `copy.deepcopy` (license included)
that falls back to `safe_repr` instead of the pickling step.
This ensures that we have a deepcopy of all data structure like objects - dicts/lists/tuples and so on
and we make a repr early for other random objects.