feat(integrations): Update StarliteIntegration to be more in line with new LitestarIntegration (#3384)
The new LitestarIntegration was initially ported from the StarliteIntegration, but then had a thorough code review that resulted in use of type comments instead of type hints (the convention used throughout the repo), more concise code in several places, and additional/updated tests. This PR backports those improvements to the StarliteIntegration. See #3358.

---------

Co-authored-by: Anton Pirker <anton.pirker@sentry.io>