Dramatiq integration from @jacobsvante (#3397)
This is the code from [sentry-dramatiq](https://github.com/jacobsvante/sentry-dramatiq). 

As described in this GitHub issue (https://github.com/getsentry/sentry-python/issues/3387) @jacobsvante,  the original maintainer of this integration, is not doing any Python anymore and wants to donate his integration to Sentry so we can take care of it. 

This PR adds the current version of the `DramatiqIntegration` to our repo. (The original integrations has been ported to the new SDK 2.x API)

Fixes #3387
---------

Co-authored-by: Ivana Kellyer <ivana.kellyer@sentry.io>