test: Make import-related tests stable
The integrations not getting enabled when there are missing modules test was relying on certain packages not being installed in the environment and was causing issues when dev requirements was installed. This patch addse a context manager that simulates import errors for certain packages to make the test robust. It also enables the redis-related test by simulating a missing 'redis' package with the same context manager.