test(starlette): Refactor shared test parametrization (#3562)
We use the same parametrization for testing FastAPI's and Starlette's `failed_request_status_codes` because the `FastApiIntegration`'s constructor is the same as `StarletteIntegration`'s constructor (the former is a subclass of the latter).

Here, we refactor the test cases to define the parametrization once, then use it in both tests. This change will make some future changes simpler, since we only need to change the parameters in one place to affect the test for both frameworks.