test(aiohttp): Delete test which depends on AIOHTTP behavior (#3568)
This test was added in #3554 to ensure that we don't break people's AIOHTTP apps when a request handler returns an invalid response. However, the test broke with a recent AIOHTTP release. After investigating, I believe the test broke because it depends on internal AIOHTTP implementation details which changed in the recent AIOHTTP release.

This test likely does not add too much value anyways, since the change in #3554 includes a comment, which explains why handling the AttributeError is important, so I think we can safely remove it.

Fixes #3567