feat: Add opportunistic Brotli compression
Brotli level 4 and 5 offer comparable or better compression to GZip
level 9 (which is our default) with better performance. This patch
adds opportunistic Brotli compression at level 4 (to be conservative)
when it detects the `brotli` module is available. It also provides some
escape hatches through `transport_compression_level` and
`transport_compression_algo` experiment configs to fine tune the behavior.

In the future, we may want to bump the default level from 4 to 5 for better
compression.