feat: Add reason for in_app to frames
Sentry enhances the in_app decision from the SDK, sometimes overriding it.

In cases the decision to set in_app to True comes from the user themselves
(by setting in_app_include), we don't want Sentry to override the decision.
Currently there's no way for Sentry to know what led to the decision in the
SDK. Changing that here.