fix(strawberry): prepare for upstream extension removal (#3649)
As suggested by @szokeasaurusrex in strawberry-graphql/strawberry#3590, Strawberry is preparing to fully remove its deprecated SentryTracingExtension in favor of the integration provided by the Sentry SDK.

This PR prepares the Sentry Strawberry integration for that removal by:
- fixing that the integration would assume Strawberry is not installed if the extension cannot be imported
- making sure tests with Strawberry versions before and after the removal still work

I also checked that removing the extension does not otherwise affect the integration: The extension's sync and async variants are imported to replace them and to guess whether sync or async code is used. Both still works if the imports are defaulted to None.