ci: Run license compliance action on all PRs (#3699)
This action only is triggered on PRs to master, but the action is required. This becomes a problem when a PR is opened against a branch other than master (e.g. as part of a PR tree). When the parent branch is merged to master, the PR's base automatically changes to master, but this action does not get triggered. Instead, it blocks on "Expected" and can only be run by adding commits to the branch.

Running the action on PRs against any branch should fix this.

Also, add logic to cancel in-progress workflows on pull requests (logic taken from our other actions)