Fix aws lambda tests (by reducing event size) (#3770)
Our AWS Lambda tests rely on outputting our events as JSON to stdout and parsing this output. AWS Lambda limits the amount of stdout it returns.

So by reducing the size of the events we can fix the tests, that where broken by printing to much data to stdout so the output is truncated and can not be parsed into actual JSON structures again.