feat(flags): add Unleash feature flagging integration (#3888)
Adds an integration for tracking flag evaluations from [Unleash](https://www.getunleash.io/) customers. 

Implementation
Unleash has no native support for evaluation hooks/listeners, unless the user opts in for each flag. Therefore we decided to patch the `is_enabled` and `get_variant` methods on the `UnleashClient` class. The methods are wrapped and the only side effect is writing to Sentry scope, so users shouldn't see any change in behavior. 

We patch one `UnleashClient` instance instead of the whole class. The reasons for this are described in
- https://github.com/getsentry/sentry-python/pull/3895

It's also safer to not modify the unleash import.

References
- https://develop.sentry.dev/sdk/expected-features/#feature-flags
- https://docs.getunleash.io/reference/sdks/python for methods we're patching/wrapping

---------

Co-authored-by: Anton Pirker <anton.pirker@sentry.io>
Co-authored-by: Colton Allen <colton.allen@sentry.io>