Centralize minimum version checking (#3910)
For [populating tox
automatically](https://github.com/getsentry/sentry-python/issues/3808),
we need to store min versions of frameworks/libraries in a
programmatically accessible place.

The obvious place for this would be in each integration; however, since
integrations can't be imported unless the respective framework is
installed, this couldn't be used from the script (unless we'd always
install all requirements of all integrations prior to running it, which
takes a non trivial amount of time). So instead I've opted for a central
place within `sentry_sdk/integrations/__init__.py`.

Note: the min versions probably need updating. Not sure when this was
last done, but some of them look quite ancient and we probably don't
support them because we'd already dropped the last Python version they'd
be able to run on.