feat(flags): remove Unleash get_variant patching code (#3914)
Follow-up to https://github.com/getsentry/sentry-python/pull/3888

The original PR patched 2 methods used for evaluating feature flags,
`is_enabled` (simple toggle on/off) and `get_variant` (returns a dict of
metadata, see
https://docs.getunleash.io/reference/sdks/python#getting-a-variant).

We want to remove all `get_variant` code since we only support boolean
flag evals atm. It seems like the main usecase for variants is reading
payloads (non-bool) for A/B/multivariate testing. This could lead to a
lot of extraneous flags, so until it is requested and/or we support
non-bool values, let's not patch this method.