Move the GraphQL group over to the tox gen script (#3975)
- remove hardcoded entries for `ariadne`, `gql`, `graphene`,
`strawberry` from the tox template
- remove them from the ignore list in `populate_tox.py`
- run `populate_tox.py` to fill in entries for them
- run `split_gh_tox_actions.py` to generate the CI yaml files so that
they correspond to the new tox.ini

Note that this effectively eliminates the `-latest` tests for the
GraphQL group. The script doesn't generate any `-latest` tests since it
always makes sure to add a pinned entry for the latest version. So in
case all of the integrations in a single group are using the script, the
whole `-latest` test category is removed.