tests: Test relevant prereleases and allow to ignore releases (#4073)
If a package has a prerelease of a higher version than the highest
released stable version, make sure to test it, too. We consider alpha,
beta, and RC releases.

Also add an option to ignore specific releases (this is related to the
above since the script now pulls in two irrelevant alpha releases of
starlite).

Closes https://github.com/getsentry/sentry-python/issues/4030

---------

Co-authored-by: Daniel Szoke <7881302+szokeasaurusrex@users.noreply.github.com>