Update tox setup for AWS Lambda integration (#4076)
Because we changed our AWS Lambda test suite to not create real Lambda
functions but rather run a Lambda environment locally, we do not need
all the code necessary for handling the authentication to AWS in the
test suite.

This PR removes all the AWS auth related code and als moves the AWS
Lambda testsuite again into the "Cloud" group of test suites, because it
is a "normal" test suite again.

The PR that includes the change to have a local Lambda environment is
here: #3988