feat(tracing): Use `sample_rand` for sampling decisions (#4044)
Use the `sample_rand` value from an incoming trace to make sampling
decisions, rather than generating a random value. When we are the head
SDK starting a new trace, save our randomly-generated value as the
`sample_rand`, and also change the random generation logic so that the
`sample_rand` is computed deterministically based on the `trace_id`.

Depends on:
  - #4040 
  - #4038 

Closes #3998