feat(tracing): Propagate sample_rand to transaction's baggage  (#4040)
`continue_trace` now propagates incoming `sample_rand` values to the
transaction's baggage. Also, in the case where `sample_rand` is missing
from the incoming trace and needs to be backfilled, this change
introduces a mechanism for the backfilled value from the scope's
propagation context to be propagated to the transaction's baggage.

The transaction still does not use the `sample_rand` for making sampling
decisions; this PR only enables propagation. A future PR will add
support for reading the incoming/backfilled `sample_rand` and for using
this value to make sampling decisions.

Depends on:
  - #4038 

Ref #3998