feat(tracing): Backfill missing `sample_rand` on `PropagationContext` (#4038)
Whenever the `PropagationContext` continues an incoming trace (i.e.
whenever the `trace_id` is set, rather than being randomly generated as
for a new trace), check if the `sample_rand` is present and valid in the
incoming DSC. If the `sample_rand` is missing, generate it
deterministically based on the `trace_id` and backfill it into the DSC
on the `PropagationContext`.
    
When generating the backfilled `sample_rand`, we ensure the generated
value is consistent with the incoming trace's sampling decision and
sample rate, if both of these are present. Otherwise, we generate a new
value in the range [0, 1).
    
Additionally, we propagate the `sample_rand` to transactions
generated with `continue_trace` (allowing the `sample_rand` to be
propagated on outgoing traces), and also allow `sample_rand` to be
used for making sampling decisions.

Ref #3998

---------

Co-authored-by: Ivana Kellyer <ivana.kellyer@sentry.io>