fix(quart): Support `quart_flask_patch` (#4132)
See
https://github.com/getsentry/sentry-python/issues/2709#issuecomment-2006932012

If `quart_flask_patch` is imported, it monkeypatches stuff so that the
Quart app appears to be a Flask app. This confuses our Flask
integration, which tries to enable itself and fails.

This commit:

- Makes the Flask integration detect that what it sees as Flask might
actually be Quart.
- Reorganizes the Quart test suite a little to allow to test this case
(a bit tricky since `import quart_flask_patch` needs to happen before
anything else due to its monkeypatching nature).

Closes https://github.com/getsentry/sentry-python/issues/2709