Reset `DedupeIntegration`'s `last-seen` if `before_send` dropped the event (#4142)
Imagine an app throws an exception twice, from different places. The
first exception is dropped in the user's `before_send`. The second
exception is not. Should the second exception appear in Sentry?

The current state is that it won't, since `DedupeIntegration` will take
the first, dropped exception into account. When encountering the second
exception, it'll consider it a duplicate and will drop it, even though
the first exception never made it to Sentry.

In this PR, we reset `DedupeIntegration`'s `last-seen` if an event has
been dropped by `before_send`, ensuring that the next exception will be
reported.

Closes https://github.com/getsentry/sentry-python/issues/371

---------

Co-authored-by: Anton Pirker <anton.pirker@sentry.io>