Better handling of exception groups (#4164)
Properly handle grouped and chained exceptions. The test case in the
linked issue illustrates that some ExceptionGroups have been handled in
a wrong way.

Updated some tests, because now that those are handled correctly all the
mechanism types except for the root exception are set to "chained" like
described in the RFC:
https://github.com/getsentry/rfcs/blob/main/text/0079-exception-groups.md#interpretation

Because this will change the grouping of exiting Sentry Issues
containing ExceptionGroups, it is safer to release this fix in the next
major and make sure that we describe the change in behavior in the
changelog. (Note: The grouping in the Ariadne issues will not change
because those are not ExceptionGroups and only updating the
`mechanism.type` does not change the grouping)

Fixes #3913