Fix AWS Lambda tests (#4199)
With the refactoring of the AWS Lambda test suite in `master`. We need
to update the tests in `potel-base` to succeed again.

This also fixes a check in the `stdlib` integration. With local AWS
Lambda we have a DSN that includes a port. This did not work and this PR
makes this work. (This will allow support for self hosted Sentry running
on a specific port too)