feat: Do not spam sentry_sdk.warnings logger w/ Spotlight (#4219)
Sometimes one may have Spotlight turned on in the SDK but not have the
sidecar running or reachable. In that case we spam the console with
every event as they fail to reach Spotlight. This patch limits the fail
warnings to 3: the first 2 are actual errors and the final one is a note
about shutting up.