fix(logging): Send raw logging parameters
This reverts commit 4c9731bbe68b6523cccec73fb764e04e61e441cb, adding tests to ensure the correct behavior going forward.

That commit caused a regression when `record.args` contains a dictionary. Because we iterate over `record.args`, that change caused us to only send the dictionary's keys, not the values.

A more robust fix for #3660 will be to send the formatted message in the [`formatted` field](https://develop.sentry.dev/sdk/data-model/event-payloads/message/) (which we have not been doing yet). I will open a follow-up PR to do this.

Fixes #4267