ref(logging): Clarify separate warnings case is for Python <3.11
The way the code was written before this change made it look like log records from the `warnings` module were always being handled by a separate code path. In fact, this separate path is only used for Python 3.10 and below.

This change makes it clear that the branch is version specific. That way, when we eventually stop supporting 3.10, it is clear that we can delete this separate block.

Depends on:
  - #4292
  - #4291