fix: Data leak in ThreadingIntegration between threads (#4281)
It is possible to leak data from started threads into the main thread
via the scopes. (Because the same scope object from the main thread
could be changed in the started thread.)
This change always makes a fork (copy) of the scopes of the main thread
before it propagates those scopes into the started thread.