test(logs): Avoid failure when running with integrations enabled
When (at least) one of integrations is enabled (because some
dependencies are installed in the environment), `sentry.sdk.name` is
changed from `sentry.python` to
`sentry.python.[FIRST_ENABLED_INTEGRATION]` which makes
`test_logs_attributes` fail. Prevent failure by relaxing the check.

This change is beneficial not only for packaging (this patch was
required for packaging for Fedora), but also for running tests with
`pytest` directly.