tests: Fix version picking in toxgen (#4323)
Toxgen should only consider the highest patch release of each
`major.minor` version. For the most part this was working fine as long
as the releases were ordered as expected in PyPI, but in cases where a
lower patch version succeeded a higher patch version in the release list
from PyPI, we would incorrectly consider the lower patch version as
well, instead of ignoring it in favor of the higher patch.

Example:
- we pull releases `[1.2.3, 1.2.4, 1.2.5, 1.2.2]` from PyPI (in that
order)
- we consolidate `1.2.3, 1.2.4, 1.2.5` into one version, `1.2.5`, as
expected
- `1.2.2` will not disappear into `1.2.5` because of a faulty check in
toxgen and will instead be considered as a new version
- our resulting list of releases eligible for testing will be `[1.2.5,
1.2.2]` instead of just `[1.2.5]`, which then results in picking
versions that are not nicely spaced apart