Introduce underscore span attributes (#4331)
Introduce the convention of underscore-prefixed span attributes. These
won't be sent to Sentry and are meant for internal SDK usage.

Changed `flag.count` to internal. Looked through the rest of the attrs
we're setting and that stuff requires a big comprehensive cleanup
altogether to make stuff align with OTel. Didn't touch anything else for
now.

Closes https://github.com/getsentry/sentry-python/issues/4329