tests: bump test timeout for recursion stacktrace extract to 2s
In some loaded environments, the test may take slightly longer than 1s
to extract the stacktrace. This was noticed in nixpkgs build system
where the load is generally high due to high build parallelism and
resource constraints. I was sometimes getting failures because the time
it took was e.g. ~1.2s (less than current timeout of 1s).

Disclosure: we'll probably end up disabling the test in nixpkgs anyway
because we try to avoid time sensitive tests. Regardless, this bump may
help someone else in a similar situation or environment.

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>