Fix FastAPI recursive exceptions (#4334)
In some cases FastAPI emits an exception that has as `__cause__` an
ExceptionGroup that contains a single excpetion. That single exepction
is the original exception. This PR prevents an infinite loop by trying
to add this construct in the `exception.values` field.

It also introduces an hard upper limit of chained/nested transaction, to
never run into an infinite loop.