fix(integrations/ray): Correctly pass keyword arguments to ray.remote function
Monkey-patched implementation was passing the provided keyword arguments incorrectly
due to a typo - "*kwargs" was used instead of "**kwargs" twice.

Fixed integration started hitting an assert in the Ray codebase that
requires for users to use "@ray.remote" decorator either with no arguments and no parentheses,
or with some of the arguments provided.

An additional wrapper function was added to support both scenarios.