test: Assert `traces_sampler` called once
Currently, this test only asserts that the `traces_sampler` is called at least once with the given arguments. However, we would expect it to be called exactly once.

This PR changes the assertion to assert that the `traces_sampler` was called exactly one time, and that that call included the given arguments.