test: Assert `traces_sampler` called once (#4450)
Currently, this test only asserts that the `traces_sampler` is called at
least once with the given arguments. However, we would expect it to be
called exactly once.

This PR changes the assertion to assert that the `traces_sampler` was
called exactly one time, and that that call included the given
arguments.

<!-- Describe your PR here -->

---

Thank you for contributing to `sentry-python`! Please add tests to
validate your changes, and lint your code using `tox -e linters`.

Running the test suite on your PR might require maintainer approval.