Patch `TracerProvider` if it already exists (#4455)
If a `TracerProvider` already exists, patch it. `TracerProvider` is a
singleton, so if we aren't the first ones setting it up, we need to use
the existing one.

In tests, reset `TracerProvider` after each test so that we start with a
clean slate and it gets set up anew on init.