ref(langchain): Greatly simplify `_wrap_configure` (#4479)
Resolving #4443 requires some changes to this method, but the current
`args`/`kwargs` business makes the method difficult to reason through.

This PR simplifies the logic by listing out the parameters we need to
access, so we don't need to access them through `args` and `kwargs`.

We also cut down on the amount of branching and the amount of variables
(`new_callbacks` vs `existing_callbacks`).

Behavior does not change in this PR; we fix the #4443 bug in #4485,
which is based on this PR
<!-- Describe your PR here -->

---

Thank you for contributing to `sentry-python`! Please add tests to
validate your changes, and lint your code using `tox -e linters`.

Running the test suite on your PR might require maintainer approval.