Support `openai-agents` (#4437)
Add support for AI agents projects using `openai-agents`
(https://pypi.org/project/openai-agents/)

Docs PR is here: https://github.com/getsentry/sentry-docs/pull/14113

This integration:
- records tracing data of agent invocation, tool execution, ai client
requests to LLMs, and handoffs to other agents.
- captures input and output to/from LLMs if `set_default_pii=True`.
- is mostly compatible to the OpenTelememetry `gen_ai` semantic
conventions. (input and output is not compatible because Sentry does not
have Span events. This information is stored in arrays on the Span
attributes.
- Captures errors that happen during agent execution (like problems
during interaction with the LLM.

This integration does not:
- Capture errors during function tool exection because this is very hard
to patch (see comment in the code)

Example span tree in Sentry.io:
![Screenshot 2025-06-24 at 12 15
17](https://github.com/user-attachments/assets/87199067-434f-4bb9-b563-5c6fc18c56cb)

---------

Co-authored-by: Ivana Kellyer <ivana.kellyer@sentry.io>