fix(tracing): Use `hasattr` for `_ctx_token`
We currently use an `if self._ctx_token` check to determine if a span has a `_ctx_token`. However, this check is incorrect, since `_ctx_token` is only sometimes set. Furthermore, when `_ctx_token` is set, we should always have a `_ctx_token`.

Therefore, using `hasattr` is what we want.