fix(profiling): Fix race condition while processing profiles
If multiple threads happen to be processing profiles, we could encounter a race condition, that would lead to more profiles being processed than are available in the `new_profiles` deque. To avoid this, we can break upon encountering an `IndexError`.

This line appears to be the most likely cause of #4542 – in any case, handling an `IndexError` is not harmful

Fixes #4542