fix(sessions): clean up threading on kill+flush
This follows the precedent set in transport.flush, and helps me with my project
to assert thread hygeine in sentry test suite (related: DI-1008), but also makes
the system much more deterministic for everyone.

This will cause shutdown to be quite slow until #4561 goes in. So I'd reject
this if you reject that.