Cleanup op and description mapping (#4560)
The main change is that `op` is now optional and `name` and
`description` are basically interchangeable and map to each other as
best as possible.

Preference is given to explicitly set values (`SentrySpanAttribute`) and
if not (in the case of pure otel spans), we try to derive them. The
entire flow is **much** cleaner now instead of repeated calls to the
same attributes.

(For the main mapping changes in `opentelemetry/test_utils` I'd
recommend reading the new code instead of the diff.)