feat(tracing): Add option to exclude specific span origins (#4463)
Allow to turn off span creation based on span origin.

This is a nice-to-have that's useful when you're double instrumenting a
library with e.g. Sentry and OpenTelemetry and you want to turn off
Sentry spans. (On the OTel side you can also disable specific
instrumentation. By adding the possibility to do this on the Sentry side
of things, we enable users to pick whichever they prefer.)