Improve token usage recording (#4566)
Update token usage recording to work if the LLM is calling them
`prompt_tokens` or `input_tokens`. Same for `completion_tokens` and
`output_tokens`. Records also cached and reasoning tokens usage.

Because the signature of a helper function was changed, other AI
integrations also have changes.

This PR does not change behavior, just prepare the ground for future
changes to the AI integrations.

---------

Co-authored-by: Ivana Kellyer <ivana.kellyer@sentry.io>