fix(clickhouse): Don't eat the generator data (#4669)
Currently, the Clickhouse integration consumes any data passed as a
generator when reading it for insertion as `db_params`. Instead, since
generators cannot be cloned, we need to wrap the generator to add the
params as we iterate over it.

Fixes #4657

<!-- Describe your PR here -->

---

Thank you for contributing to `sentry-python`! Please add tests to
validate your changes, and lint your code using `tox -e linters`.

Running the test suite on your PR might require maintainer approval.