Change Lambda timeout warning message (#4671)
The timeout in the AWS Lambda (and GCP, by extension) timeout warning is
wrong, remove it and slightly rephrase.

This is on `potel-base` since it breaks grouping. (Even just removing
the number from the original message, without modifying it further,
causes regrouping. Since that's the case I took the liberty to update
the message a bit more.)

Fixes https://github.com/getsentry/sentry-python/issues/4629