Add transport worker factory and abstract base class for background worker (#4580)
Add a new abstract base class for the background worker implementations
in the transport. This allows for implementation of the upcoming async
task-based worker in addition to the current thread based worker used in
the sync context. Additionally, add a factory method in the
HttpTransportCore shared class, to allow the worker methods to live
higher up in the class hierarchy regardless of specific implementation.

GH-4578