Add async task background worker (#4591)
Add a new implementation of the transport background worker based on an
async task. This worker mostly mirrors the same functionality as the
thread-based worker, with the exception that it exposes a non-blocking
async flush (which can be awaited from an async context). Furthermore,
the worker itself is not thread-safe and should be called using
[run_coroutine_threadsafe](https://docs.python.org/3/library/asyncio-task.html#asyncio.run_coroutine_threadsafe)
or similar when called from another thread (this is fixed and handled by
the transport). I have kept the fork check from the threaded worker, but
I am not sure if it is necessary as forking in an async application
would also break the event loop.

GH-4581

---------

Co-authored-by: Neel Shah <neel.shah@sentry.io>